Revert "Refactor : CustomUserDetails DTO로 이동"
This reverts commit 3800e73204982a47ef6a4243c88bd2f721b746d3.
